feat: добавлена поддержка GroupVisitObjectId и CategoryVisitorId в DTO объектов посещения

+++ src/main/java/com/ticket/terminal/dto/VisitObjectDto.java
@@ -25,7 +25,7 @@ public class VisitObjectDto {
     private Long groupVisitObjectId;
 
     @JsonProperty("CategoryVisitorId")
-    private Integer categoryVisitorId;
+    private Long categoryVisitorId;
 
     @JsonProperty("Address")
     private String address;
